+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Display "in group {group}" subheader if post is addressed to a group.
 - Added "Group members" page.
 
+### Changed
+
+- Updated translations ([#174](https://codeberg.org/silverpill/mitra-web/pulls/174)).
+
 ## [5.7.1] - 2026-07-12
 
 ### Changed