+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [5.8.0] - 2026-07-31
+
 ### Added
 
 - Added "Reactions" tab to user's own profile page.